improved project structure and assigned componenet prop types
